feat: add "--no-validate" flag
